+++ test/TestSuite.java
@@ -14,7 +14,10 @@ import org.junit.BeforeClass;
 	StackTest.class,
 	QueueTest.class,
 	DynamicArrayTest.class,
-	SimpleSortsTest.class,
+	SortTest.class,
+	InsertionSortTest.class,
+	SelectionSortTest.class,
+	BubbleSortTest.class,
 	MergeSortTest.class,
 	QuicksortTest.class,
 	HeapSortTest.class,
